Improved the build process for the examples.

+++ backend/update_examples.sh
@@ -7,16 +7,18 @@ EXAMPLES_SYNT_FOLDER=./examples/synthesis/forallsafety/
 
 ###### Model checking
 echo "%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% MODEL CHECKING"
-echo "> MC: Remove all current examples ..."
-rm -r $MC_FOLDER
-mkdir $MC_FOLDER
+if [[ ! -z ${clean} ]]; then
+    echo "> MC: Remove all current examples ..."
+    rm -r $MC_FOLDER
+    mkdir $MC_FOLDER
+fi
 echo "> MC: Copy examples with coordinates..."
 for file in $(find $EXAMPLES_MC_FOLDER -name '*.apt'); do   # all .apt files recursively
     if grep -q "xCoord" "$file"; then
         # tinker the new location
         new_file=${file/"$EXAMPLES_MC_FOLDER"/"$MC_FOLDER"}
         new_folder=${new_file%/*}"/"
-        mkdir $new_folder   # create the folder 
+        mkdir -p $new_folder   # create the folder 
         echo $file" -> "$new_file
         cp $file $new_file  # copy the file
     fi
@@ -24,9 +26,11 @@ done
 
 ###### Synthesis
 echo -e "\n%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% SYNTHESIS"
-echo "> SYNT: Remove all current examples ..."
-rm -r $SYNT_FOLDER
-mkdir $SYNT_FOLDER
+if [[ ! -z ${clean} ]]; then
+    echo "> SYNT: Remove all current examples ..."
+    rm -r $SYNT_FOLDER
+    mkdir $SYNT_FOLDER
+fi
 echo "> SYNT: Copy examples with coordinates..."
 for file in $(find $EXAMPLES_SYNT_FOLDER -name '*.apt'); do   # all .apt files recursively
     if grep -q "xCoord" "$file"; then
